services/speak/requirements.txt
PyYAML>=6.0

fastapi>=0.110
uvicorn[standard]>=0.27
